requirements.txt
feedparser==6.0.10
pyyaml==6.0.1
python-dateutil==2.8.2
tabulate==0.9.0